Make VariableList a "core" type.
