+++ opencog/atoms/core/CMakeLists.txt
@@ -5,6 +5,7 @@ INCLUDE_DIRECTORIES( ${CMAKE_CURRENT_BINARY_DIR})
 ADD_LIBRARY (atomcore SHARED
 	FreeLink.cc
 	PutLink.cc
+	VariableList.cc
 )
 
 # Without this, parallel make will race and crap up the generated files.
@@ -21,5 +22,6 @@ INSTALL (TARGETS atomcore
 INSTALL (FILES
 	FreeLink.h
 	PutLink.h
+	VariableList.h
 	DESTINATION "include/opencog/atoms/core"
 )
