+++ opencog/atoms/core/VariableList.cc
