+++ opencog/atomspace/AtomTable.cc
@@ -43,12 +43,12 @@
 #include <opencog/atoms/bind/DefineLink.h>
 #include <opencog/atoms/bind/LambdaLink.h>
 #include <opencog/atoms/bind/PatternLink.h>
-#include <opencog/atoms/bind/VariableList.h>
+#include <opencog/atoms/core/PutLink.h>
+#include <opencog/atoms/core/VariableList.h>
 #include <opencog/atoms/execution/EvaluationLink.h>
 #include <opencog/atoms/execution/ExecutionOutputLink.h>
 #include <opencog/atoms/reduct/DeleteLink.h>
 #include <opencog/atoms/reduct/FunctionLink.h>
-#include <opencog/atoms/core/PutLink.h>
 #include <opencog/util/exceptions.h>
 #include <opencog/util/functional.h>
 #include <opencog/util/Logger.h>
